Forgot Password and Change Password feature
Introduced both forgot and change password features to allow users the option of setting new passwords. As a result, additions were made to the unit test file to include testing for said features. Also, the user is now unable to access other pages within the application unless they are signed in and are now able to logout of their account. One issue does exist in which the change password is unreachable since it is on the login page but once logged in you cant access that page. We should look for another place to include it.