Change Password Fixed
Removed the Change Password feature from the login page to the profile page. If a user wanted to change their password they could go under Account details in the profile page where they will be directed to a new page that displays their email and the Change Password button. Also changed the text that is displayed when successfully registering an account so that only what is wanted to be display is included and none of the raw text is displayed.